230818 BE Merge to Main

+++ backend/src/main/java/com/vvs/peekpick/peek/dto/PeekNearSearchDto.java
@@ -9,5 +9,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class PeekNearSearchDto {
     String peekId;
-    Double distance;
+    int distance;
 }
